+++ src/main/java/org/idnt/udemy/mockitoapp/example/repository/ExamRepository.java
@@ -1,0 +1,11 @@
+package org.idnt.udemy.mockitoapp.example.repository;
+
+import org.idnt.udemy.mockitoapp.example.model.Exam;
+
+import java.util.List;
+
+public interface ExamRepository {
+    List<Exam> findAll();
+
+    Exam save(Exam exam);
+}